requirements.txt
einops
hydra-core
numpy
omegaconf
lightning
wandb
pytest
pandas
matplotlib